Added expression-util and removed unused util functions
